Refactor the project structure
