+++ src/layout/SharedLayout.jsx
@@ -1,6 +1,6 @@
 import { Outlet } from "react-router-dom";
-import Header from "./Header";
-import Aside from "./Aside";
+import Header from "../layout/Header";
+import Aside from "../layout/Aside";
 
 export default function SharedLayout() {
   return (
